PG: run docker in simpler host-networking dev mode
docker-compose isn't apparently compatible with my version of ip-tables
or something and ends up failing to run the needed ip-tables command.

Specifically, the version of iptables available to me has renamed the
dest port argument name.

Host-mode networking is simpler and it just works.

Add task for running the migrations.